Fixing the management of file descriptors (sockets & files) & resolve race conditions (#47)
* [source] Fix leaked socket FD for the client socket
- Close the socket connecting the controller with the client(s) when the client sends the exit message
- Apply the fix in both the native and gdpr controller

Signed-off-by: dimstav23 <dimstav23@gmail.com>

* [CMakeLists.txt] add TSan compile option

Signed-off-by: dimstav23 <dimstav23@gmail.com>

* [README.md] adapt ASan compile option & add TSan compile option in the README

Signed-off-by: dimstav23 <dimstav23@gmail.com>

* [controller] Implement LRU policy for the file descriptors -- pending: performance testing

Signed-off-by: dimstav23 <dimstav23@gmail.com>

* [controller/source/logging] Simplify eviction policy -- remove the file stream that was opened first

Signed-off-by: dimstav23 <dimstav23@gmail.com>

* [controller/source] add better way to get the ulimit -n programmatically and some minor formatting changes to satisfy clang

Signed-off-by: dimstav23 <dimstav23@gmail.com>

* [controller/source] set SO_REUSEADDR socket property for the client socket to avoid binding issues in the testing automation

Signed-off-by: dimstav23 <dimstav23@gmail.com>

* [controller/source/logging] fix concurrency bug for logging lock acquisition
Previously, the unordered_map that holds the locks per key-file was not initialized in a thread-safe manner
leading to concurrency bugs (e.g. segfaults) in case that two threads try to initialize the same element in the unordered_map.
This commit adds a check in the get_or_open_log_stream function so that the mutex for a key-log file is initialized protected by
the global fd_mgmt mutex. In addition, after the log acquisition in the log_encoded_entry function, we add a check for a potential
race condition that the retrieved file is closed after its retrieval but before the thread acquires the log for the file.

Signed-off-by: dimstav23 <dimstav23@gmail.com>

---------

Signed-off-by: dimstav23 <dimstav23@gmail.com>